Update ruby to 2.5.3
The version before was using the Dockerfile from
https://github.com/docker-library/ruby/blob/699a04311386ecc98ca242fc9bdee17fb4008863/2.4/alpine3.7/Dockerfile
but changing only the ruby version from 2.4.4 to 2.4.1 and updating the
SHA. There is no Dockerfile for alpine3.7 for ruby 2.4.1, which I assume
is why we custom built one. Since the basic ruby install was unchanged,
I've reverted back to specifying the ruby Dockerfile.